requirements.txt
Django>=2.2,<4
drf-yasg==1.20.0
drf-nested-routers==0.91
requests==2.22.0

# API
djangorestframework==3.12.2
pandas==1.1.5
xlrd==1.2.0
